requirements.txt
Flask==1.0.2
gevent==1.3.5
gunicorn==19.9.0
h5py==2.8.0
Keras==2.2.0
numpy==1.14.5
pillow==5.2.0
tensorflow==1.9.0
Circus==0.15.0

